+++ CHANGES.md
@@ -3,12 +3,18 @@
 
 ## To-Do
 ## -----
-1)  Make a 'summary' command to run 'sacct' and summarize the results.  For example: the number of jobs in each 'state', and the typical time for completed and still running jobs, etc.
 
 
 ## Current
 ## -------
-
+-   `slurpy/__main__.py`
+    -   Added `--verbose` command-line argument to modify results.
+    -   Added `--summary` command-line argument to summarize all jobs reported by 'sacct'.  Currently returns the counts and optionally the min, max, median elapsed time for each state (if the `--verbose` flag is passed).
+-   `slurpy/sacct.py`
+    -   `summary()`
+        -   API method to parse 'sacct' command results and summarize succinctly.
+        -   Reports number of jobs in each 'state'
+        -   If `verbose`, then reports min, max, median elapsed time for each state.
 
 ## v0.1 - 2016/12/13
 ## -----------------
@@ -17,7 +23,7 @@
 -   `slurpy/const.py`
     -   File to contain constants and parameters.
 -   `slurpy/sacct.py`
-    -   File to handle interactions with the `sacct` command.
+    -   File to handle interactions with the 'sacct' command.
     -   `sacct()`
         -   API Method to call and report from the SLURM 'sacct' command.
     -   Includes methods to parse sacct results (`_parse_sacct` and `_parse_sacct_line`), filter the resulting lines based on 'state' (`_filter_lines`) and nicely format the results for printing (`_print_lines_dicts`).